test(collector): share the metric-gathering test helpers
The three helpers introduced with the drain timestamp tests are not specific
to that collector: stubbing Execute, rendering a gathered family as sorted
`name{labels} value` lines, and converting a Slurm timestamp the way the
command produced it apply to any collector test.

Move them to helpers_test.go under neutral names before a second collector
test needs them, and add bufferLogger for tests that assert a collector
reported a problem rather than staying silent. The exported embedded field on
logger.Logger makes that composable without touching the logger package.

No behaviour change: same assertions, same fixtures.
